requirements.txt
aiogram==3.21.0
python-dotenv
aiofiles
Kurigram==2.2.6
TgCrypto
aiohttp
aiohttp-socks
motor==3.6.0